Code restructuring
