+++ alphaswarm/core/prompt/base.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+import abc
+
+from pydantic import BaseModel
+
+
+class PromptTemplateBase(BaseModel, abc.ABC):
+    @abc.abstractmethod
+    def get_template(self) -> str:
+        pass
+
+# TODO: base prompt pair
